Make graph_project a common pytest fixture (move it to conftest.py)
